third_party/WebKit/LayoutTests/crypto/unwrapKey-lacks-usage-expected.txt

CONSOLE ERROR: key.usages does not permit this operation
Tests that unwrapping keys must have the 'unwrapKey' usage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
